Fixed: No prototype needed in this context.
The purpose of prototypes is so that main() can be viewed at the top of the file without having to scroll through a bunch of function definitions to get to it. The prototypes define the functions before main() so that they are not undefined when the compiler reaches their calls. You don't need to prototype a function right before defining it in a head.